+++ src/services/file/bad-script-permission.error.js
